requirements.txt
Django==4.2.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
gunicorn==21.2.0
python-decouple==3.8
yfinance==0.2.65
matplotlib==3.7.2
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0